requirements.txt
playwright==1.32.0
pytest==7.2.2


allure-pytest==2.13.1

pytest-xdist==3.2.1